% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_digits.R
\name{count_digits}
\alias{count_digits}
\title{Number of Non-decimal Digits}
\usage{
count_digits(number)
}
\arguments{
\item{number}{input desired number to count digits.}
}
\value{
count of non-decimal digits
}
\description{
This function counts the number of non-decimal digits that a number has.
}
\examples{
count_digits(55.37)
count_digits(1001.2)
}
