% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_wl_final_reserves.R
\name{discrete_wl_final_reserves}
\alias{discrete_wl_final_reserves}
\title{Discrete Whole Life Insurance Final Reserve Calculator}
\usage{
discrete_wl_final_reserves(prem = 5000, interest, mortality,
  number_alive = 10000, benefit = 4e+05, seed_deaths = 462,
  seed_rates = 873, start_age = 50)
}
\arguments{
\item{prem}{is the premium paid by each policyholder that is alive at the start of each year (default = 5000).}

\item{interest}{should be a vector of interest rates expressed as decimals.}

\item{mortality}{should be a vector of mortality rates given as decimals starting from birth.}

\item{number_alive}{is the number of policyholders alive when the life insurance first took effect.}

\item{benefit}{is the death benefit paid at the end of the year to each policyholder who dies.}

\item{seed_deaths}{is a seed for the normal distribution used when generating the number of policyholders who die each year (default = 462).}

\item{seed_rates}{is a seed for the normal distribution used when generating an interest rate for each year (default = 873).}

\item{start_age}{is the age of policyholders at when the life insurance first takes effect.}
}
\value{
The final reserve will be returned at the time when the last policyholder dies.
}
\description{
This function calculates the reserves at the time when the last policyholder dies for discrete whole life insurance that pays death benefits at the end of the year of death.
}
\examples{
rates <- c(0.01, 0.02, 0.025, 0.021, 0.03)
mortality_probs_young <- seq(from = 0.0002, to = 0.01, length.out = 61)
mortality_probs_middle_age <- seq(from = 0.011, to = 0.06, length.out = 20)
mortality_probs_old <- seq(from = 0.065, to = 0.165, length.out = 10)
mortality_probs_older <- seq(from = 0.17, to = 0.85, length.out = 30)
mortality_probs <- c(mortality_probs_young, mortality_probs_middle_age, mortality_probs_old, mortality_probs_older)
discrete_wl_final_reserves(prem = 10000, interest = rates, mortality = mortality_probs, start_age = 40)
}
