% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_later_value.R
\name{find_later_value}
\alias{find_later_value}
\title{Parameter Estimator (Besides First Parameter)}
\usage{
find_later_value(f, params, val = 0, maximum = 1e+05, minimum = 0,
  only_integers = F, decimals = 4)
}
\arguments{
\item{f}{the function that is being used (note f should be a monotonic function for this function to work properly).}

\item{params}{a list of parameter values for all parameters that are specified in a function call prior to the desired parameter as well as any required parameters that do not have default values that occur after the desired parameter.}

\item{val}{the desired value for the function.}

\item{maximum}{the maximum value for the search parameter (default = 100000).}

\item{minimum}{the minimum value for the search parameter (default = 0).}

\item{only_integers}{logical indicating whether the function only takes integers for the search parameter (default = F, allowing decimals).}

\item{decimals}{the number of decimal places to use for the search parameter.}
}
\value{
A list will be returned containing the estimate for the desired parameter, the function value when the parameter estimate is used in f, and a warning if the parameter estimate results in a function value that is more than 1 away from val.
}
\description{
This function finds the value of a parameter other than the first for a function that gives a desired value.
}
\examples{
find_later_value(log, params = list(32), val = 2, decimals = 6)
}
